+++ lib/mir/src/ir.rs
@@ -9,13 +9,13 @@ mod case;
 mod clone_variables;
 mod comparison_operation;
 mod comparison_operator;
-mod declaration;
 mod default_alternative;
-mod definition;
 mod drop_variables;
 mod expression;
 mod foreign_declaration;
 mod foreign_definition;
+mod function_declaration;
+mod function_definition;
 mod if_;
 mod let_;
 mod let_recursive;
@@ -38,13 +38,13 @@ pub use case::*;
 pub use clone_variables::*;
 pub use comparison_operation::*;
 pub use comparison_operator::*;
-pub use declaration::*;
 pub use default_alternative::*;
-pub use definition::*;
 pub use drop_variables::*;
 pub use expression::*;
 pub use foreign_declaration::*;
 pub use foreign_definition::*;
+pub use function_declaration::*;
+pub use function_definition::*;
 pub use if_::*;
 pub use let_::*;
 pub use let_recursive::*;
